// confmod-app/frontend/src/app/openapi/model/rawConfmodModelInput.ts
/**
 * Confmod Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RawConfmodHeaderInput } from './rawConfmodHeaderInput';
import { RawConfmodModelInputPayloadInnerInner } from './rawConfmodModelInputPayloadInnerInner';


export interface RawConfmodModelInput { 
    head: RawConfmodHeaderInput;
    payload: Array<Array<RawConfmodModelInputPayloadInnerInner>>;
}